fixed pep8 warnings, typos, for loop in basket.html

+++ products/views.py
@@ -36,15 +36,21 @@ def basket_add(request, product_id):
 
     return HttpResponseRedirect(request.META['HTTP_REFERER'])
 
+
 def basket_view(request):
     context = {
         'baskets': Basket.objects.filter(user=request.user)
     }
     return render(request, 'products/basket.html', context)
 
+
 def basket_remove(request, id):
     basket = Basket.objects.get(pk=id)
-    basket.delete()
+    if basket.quantity == 1:
+        basket.delete()
+    else:
+        basket.quantity -= 1
+        basket.save()
     return HttpResponseRedirect(request.META['HTTP_REFERER'])
